requirements.txt
acme==0.19.0
asn1crypto==0.23.0
certbot==0.19.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
ConfigArgParse==0.12.0
configobj==5.0.6
cryptography==2.1.1
Django==1.11.6
django-bleach
django-braces==1.11.0
django-geojson==2.10.0
django-leaflet==0.22.0
django-model-utils==3.0.0
djangorestframework==3.6.2
django-star-ratings==0.5.6
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
idna==2.6
ipaddress==1.0.18
mock==2.0.0
ndg-httpsclient==0.4.3
parsedatetime==2.4
pbr==3.1.1
psycopg2==2.7.3.1
pusher==1.7.2
pyasn1==0.3.7
pycparser==2.18
pyOpenSSL==17.3.0
pyRFC3339==1.0
pysqlite==2.8.3
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
zope.component==4.4.1
zope.event==4.3.0
zope.interface==4.4.3
